backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.57.0
twilio>=9.0.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
